+++ src/routes/profileview/components/ProfileView.tsx
@@ -3,16 +3,13 @@ import { Layout, Row, Col, Tabs, Popover, Spin } from 'antd';
 
 import Navbar from "../../../../src/Components/Shared/Navbar/NavbarContainer"
 import SidebarView from "./SidebarView";
-import Collaborators from "./Collaborators";
-import TabPaneView from "./TabPaneView";
-import UserInformationView from "./UserInformationView";
-import ProfileHeader from "./ProfileHeader";
-import ProfileFilter from "./ProfileFilter";
+
 
 
 import { useMapDispatch, useMapState } from "../../../../src/hook/mapHook";
 import { useProfileState } from "../../../../src/hook/profileHook";
 import { useAppUserDispatch } from "../../../../src/hook/useAppUser";
+import ProfileBody from "./ProfileBody";
 
 const content = (<div className="popoveer-00">View Problems within the Area of Interest, as defined by the "Default Map Zoom Area" in the "Edit Profile" section</div>);
 const content00 = (<div className="popoveer-00">View Projects within the Area of Interest, as defined by the "Default Map Zoom Area" in the "Edit Profile" section</div>);
@@ -41,42 +38,7 @@ const ProfileView = () => {
       <Layout>
         <SidebarView></SidebarView>
         <Layout className="map-00 profile-00">
-          {/* <ProfileHeader /> */}
-          <Row className="profile-header">
-              <UserInformationView
-                key="userProfile"
-                setFilter={setFilter} />
-            </Row>
-          {/* <ProfileFilter /> */}
-          <Row>
-            <Col className="profile-tabs" xs={{ span: 24 }} lg={{ span: 17 }}>
-              <Spin className="loading-01" spinning={favoritesLoader > 0}>
-              <Tabs defaultActiveKey="2" className="tabs-map" onTabClick={(e: string) => {
-                if (e == '1') {
-                  favoriteCards(user.email, true, { keyword: "", column: 'problemname', order: "asc"});
-                } else {
-                  favoriteCards(user.email, false, { keyword: "", column: 'projectname', order: "asc"});
-                }
-              }} style={{marginLeft: '24px'}}>
-                <TabPane key="1" tab={<span><Popover content={content} placement="rightBottom">Problems</Popover> </span>} >
-                  <TabPaneView
-                    type={"Problems"}
-                    data={favoriteProblemCards}
-                    filter={filter}
-                  />
-                </TabPane>
-                <TabPane key="2" tab={<span><Popover content={content00} placement="rightBottom">Projects</Popover> </span>}>
-                  <TabPaneView
-                    type={"Projects"}
-                    data={favoriteProjectCards}
-                    filter={filter}
-                  />
-                </TabPane>
-              </Tabs>
-              </Spin>
-            </Col>
-            <Collaborators />
-          </Row>
+          <ProfileBody />
         </Layout>
       </Layout>
     </Layout>
