+++ src/routes/profileview/components/SidebarModal.tsx
@@ -1,7 +1,5 @@
 import React, { useState } from "react";
 import { Modal } from 'antd';
-import { useLocation } from "react-router-dom";
-import store from "../../../store";
 import '../../../Scss/Components/sidebar.scss';
 
 const stateValue = {
@@ -10,11 +8,6 @@ const stateValue = {
 
 export default () => {
     const [state, setState] = useState(stateValue);
-    const showModal = () => {
-      const auxState = {...state};
-      auxState.visible = true;
-      setState(auxState);
-    };
 
     const handleOk = (e: any) => {
       console.log(e);
@@ -29,14 +22,6 @@ export default () => {
       auxState.visible = false;
       setState(auxState);
     };
-
-    //  const [collapsed, setCollapsed] = useState<boolean>(true);
-     const location = useLocation();
-     const appUser = store.getState().appUser;
-    //  const indexOf = "" + ROUTERS_SIDEBAR.indexOf(location.pathname);
-
-    //  const showWorkRequestPlan = (appUser.designation !== 'guest' && appUser.designation === 'admin' || appUser.designation === 'staff' || appUser.designation === 'government_staff')
-    //  const userApproved = appUser.status === 'approved';
 
   return (
     <Modal
