+++ src/routes/profileview/constants/layout.constants.ts
@@ -2290,16 +2290,4 @@ export const ICON_POPUPS = [
   ['Stream Management Corridor', "/Icons/ic_SMC_Watershed@2x.png"],
 ]
 
-/*  // Mapbox Custom Styles
-    mapbox://styles/mapbox/streets-v11
-    mapbox://styles/mapbox/outdoors-v11
-    mapbox://styles/mapbox/light-v10
-    mapbox://styles/mapbox/dark-v10
-    mapbox://styles/mapbox/satellite-v9
-    mapbox://styles/mapbox/satellite-streets-v11
-    mapbox://styles/mapbox/navigation-preview-day-v4
-    mapbox://styles/mapbox/navigation-preview-night-v4
-    mapbox://styles/mapbox/navigation-guidance-day-v4
-    mapbox://styles/mapbox/navigation-guidance-night-v4
-*/
